Add frontend for posts
