+++ frontend/components/post.js
@@ -1,8 +1,7 @@
-import React from "react";
-import Link from "next/link";
+import React from 'react'
+import Link from 'next/link'
 
-import "../public/stylesheets/materialize.min.css";
-import "./post.css";
+import './post.css'
 
 /*
     title: Indicates title of post
@@ -15,45 +14,47 @@ import "./post.css";
 */
 
 export const Post = ({
-	//Default Values
-	postid = 0,
-	title = "Title Missing",
-	author = "Author Missing",
-	date = "Date Missing",
-	details = "Details Missing",
-	categ = ["#something1", "Something2"]
+  // Default Values
+  postid = 0,
+  title = 'Title Missing',
+  author = 'Author Missing',
+  date = 'Date Missing',
+  details = 'Details Missing',
+  categ = ['#something1', 'Something2']
 }) => {
-	return (
-		<div className="row post">
-			<div className="col s12">
-				<div className="card hoverable blue-grey darken-1">
-					<div className="card-content white-text">
-						<span className="card-title">{title}</span>
-						<blockquote className="subtitle">
-							<p>
-								By: {author} on {date}
-							</p>
-						</blockquote>
+  return (
+    <div className="row post">
+      <div className="col s12">
+        <div className="card hoverable blue-grey darken-1">
+          <div className="card-content white-text">
+            <span className="card-title">{title}</span>
+            <blockquote className="subtitle">
+              <p>
+                By:
+                {author}
+                on
+                {date}
+              </p>
+            </blockquote>
 
-						<p>{details}</p>
-					</div>
-					<div className="tags">
-						{categ.map((cat, i) => {
-							return (
-								<div key={i} className="chip">
-									{cat}
-								</div>
-							);
-						})}
-					</div>
-					<div className="card-action">
-						<a href="#">Read more</a>
-						<Link href={`/post?postid=${postid}`}>
-							<a>Read More</a>
-						</Link>
-					</div>
-				</div>
-			</div>
-		</div>
-	);
-};
+            <p>{details}</p>
+          </div>
+          <div className="tags">
+            {categ.map((cat, i) => {
+              return (
+                <div key={i} className="chip">
+                  {cat}
+                </div>
+              )
+            })}
+          </div>
+          <div className="card-action">
+            <Link href={`/posts/[id]`} as={`/posts/${postid}`}>
+              <a>Read More</a>
+            </Link>
+          </div>
+        </div>
+      </div>
+    </div>
+  )
+}
