+++ frontend/pages/_app.js
@@ -1,0 +1,29 @@
+import App, { Container } from 'next/app'
+import React from 'react'
+import Layout from '../components/Layout'
+
+class MyApp extends App {
+  static async getInitialProps({ Component, ctx }) {
+    let pageProps = {}
+
+    if (Component.getInitialProps) {
+      pageProps = await Component.getInitialProps(ctx)
+    }
+
+    return { pageProps }
+  }
+
+  render() {
+    const { Component, pageProps } = this.props
+
+    return (
+      <Container>
+        <Layout>
+          <Component {...pageProps} />
+        </Layout>
+      </Container>
+    )
+  }
+}
+
+export default MyApp
